requirements.txt
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
ibm-watsonx-orchestrate>=1.14.0
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
Flask==3.0.0
redis==5.0.1
pytest==7.4.3
pytest-cov==4.1.0
pyyaml==6.0.1
qrcode==7.4.2
pillow==10.1.0
numpy==1.26.2
